refactor: Rename and reorganize WebSocket-related modules and remove unused socket file
