+++ src/chart/history/mod.rs
@@ -4,7 +4,7 @@ pub mod batch;
 pub mod single;
 
 /// Resolve authentication token from parameter or environment
-fn resolve_auth_token(auth_token: Option<&str>) -> Result<String> {
+fn _resolve_auth_token(auth_token: Option<&str>) -> Result<String> {
     match auth_token {
         Some(token) => Ok(token.to_string()),
         None => {
