+++ src/utils.rs
@@ -1,6 +1,6 @@
 use crate::{
     Result, UserCookies,
-    live::socket::{SocketMessage, SocketMessageDe},
+    live::models::{SocketMessage, SocketMessageDe},
     models::{MarketAdjustment, SessionType},
 };
 use base64::engine::{Engine as _, general_purpose::STANDARD as BASE64};
